add profiler_help.h to refine the code test=develop
